Refactorisation EditProfileUi.cs : extraction du module Moodles en composant `MoodlesEditor` (state, IPC fetch/apply/restore, UI complète).
