+++ UmbraSync/UI/Components/StatusIconInfo.cs
@@ -1,0 +1,3 @@
+namespace UmbraSync.UI.Components;
+
+public readonly record struct StatusIconInfo(uint IconId, string Name);
